fix: address PR review comments
- Move clearCompatibilityCaches (and its LRUCache-internals workaround)
  out of production code into a test-only module
  (compatibilityCacheTestUtils.ts), per review feedback that it's only
  ever used by tests.
- Consolidate node/resolvers/stats/constants.ts into
  node/resolvers/search/constants.ts — PATH_SEPARATOR/MAP_SEPARATOR
  there duplicated PATH_SEPARATOR/MAP_VALUES_SEP already defined in
  search/constants.ts, and newURLs.ts was the only consumer.
- Add a comment on resolvers/stats/index.ts noting it's kept only
  because vtex.search-graphql's schema still declares searchURLsCount,
  so it isn't deleted without a coordinated schema cleanup.
- Fix a grammar nit in a test description.
