+++ node/resolvers/search/modules/compatibilityCacheTestUtils.test.ts
@@ -4,8 +4,8 @@ import {
   categoryTreeRootCache,
   searchUrlsCache,
   facetsCache,
-  clearCompatibilityCaches,
 } from './compatibilityCache'
+import { clearCompatibilityCaches } from './compatibilityCacheTestUtils'
 
 describe('clearCompatibilityCaches', () => {
   it('empties every compatibility cache', async () => {
@@ -17,13 +17,15 @@ describe('clearCompatibilityCaches', () => {
     ]
 
     await Promise.all(
-      caches.map(cache => getOrSet(cache, 'some-key', async () => 'some-value'))
+      caches.map((cache) =>
+        getOrSet(cache, 'some-key', async () => 'some-value')
+      )
     )
 
-    caches.forEach(cache => expect(cache.has('some-key')).toBe(true))
+    caches.forEach((cache) => expect(cache.has('some-key')).toBe(true))
 
     clearCompatibilityCaches()
 
-    caches.forEach(cache => expect(cache.has('some-key')).toBe(false))
+    caches.forEach((cache) => expect(cache.has('some-key')).toBe(false))
   })
 })
