--- node/resolvers/stats/constants.ts
@@ -1,2 +1,0 @@
-export const PATH_SEPARATOR = '/'
-export const MAP_SEPARATOR = ','
